работающая версия опросника для второго урока
