--- hw/src/main/java/ru/otus/lesson1/dao/PersonDaoI.java
@@ -1,7 +1,0 @@
-package ru.otus.lesson1.dao;
-
-import ru.otus.lesson1.domain.Person;
-
-public interface PersonDaoI {
-    Person getPeson();
-}
