--- hw/src/main/java/ru/otus/lesson1/view/Responsable.java
@@ -1,9 +1,0 @@
-package ru.otus.lesson1.view;
-
-import ru.otus.lesson1.domain.Person;
-import ru.otus.lesson1.domain.Question;
-import java.util.HashSet;
-
-public interface Responsable {
-    Person getAnswers(HashSet<Question> set);
-}
