+++ hw/src/main/java/ru/otus/lesson2/dao/PersonDaoI.java
@@ -1,0 +1,7 @@
+package ru.otus.lesson2.dao;
+
+import ru.otus.lesson2.domain.Person;
+
+public interface PersonDaoI {
+    Person getPeson();
+}
